--- a/src/java/picard/analysis/directed/CollectTargetedPcrMetrics.java
+++ b/src/java/picard/analysis/directed/CollectTargetedPcrMetrics.java
@@ -82,6 +82,6 @@ public class CollectTargetedPcrMetrics extends CollectTargetedMetrics<TargetedPc
                                                         final String probeSetName,
                                                         final int nearProbeDistance) {
         return new TargetedPcrMetricsCollector(accumulationLevels, samRgRecords, refFile, perTargetCoverage, targetIntervals, probeIntervals, probeSetName, nearProbeDistance,
-                MINIMUM_MAPPING_QUALITY, MINIMUM_BASE_QUALITY, CLIP_OVERLAPPING_READS, true);
+                MINIMUM_MAPPING_QUALITY, MINIMUM_BASE_QUALITY, CLIP_OVERLAPPING_READS, true, COVERAGE_CAP, SAMPLE_SIZE);
     }
 }