--- a/src/java/picard/analysis/directed/HsMetrics.java
+++ b/src/java/picard/analysis/directed/HsMetrics.java
@@ -231,4 +231,10 @@ public class HsMetrics extends MultilevelMetrics {
      * reads that should have mapped to GC>=50% regions mapped elsewhere.
      */
     public double GC_DROPOUT;
+
+    /** The theoretical HET SNP sensitivity. */
+    public double HET_SNP_SENSITIVITY;
+
+    /** The Phred Scaled Q Score of the theoretical HET SNP sensitivity. */
+    public double HET_SNP_Q;
 }