--- a/src/java/picard/util/MathUtil.java
+++ b/src/java/picard/util/MathUtil.java
@@ -134,6 +134,28 @@ final public class MathUtil {
         return index;
     }
 
+    /** Returns the largest value stored in the array. */
+    public static long max(final long[] nums) {
+        return nums[indexOfMax(nums)];
+    }
+
+    /**
+     * Returns the index of the largest element in the array.  If there are multiple equal maxima then
+     * the earliest one in the array is returned.
+     */
+    public static int indexOfMax(final long[] nums) {
+        double max = nums[0];
+        int index  = 0;
+        for (int i = 1; i < nums.length; ++i) {
+            if (nums[i] > max) {
+                max = nums[i];
+                index = i;
+            }
+        }
+
+        return index;
+    }
+
     /** Returns the smallest value stored in the array. */
     public static double min(final double[] nums) {
         double min = nums[0];