--- a/testdata/picard/sam/CollectGcBiasMetrics/MSmallHeader.dict
+++ b/testdata/picard/sam/CollectGcBiasMetrics/MSmallHeader.dict
@@ -1,0 +1,2 @@
+@HD	VN:1.0	SO:coordinate
+@SQ	SN:chrM	LN:16571	UR:chrM.reference.fasta