… src/java/picard/analysis/directed/TargetMetricsCollector.java
… src/tests/java/picard/analysis/CollectWgsMetricsTest.java
… src/tests/java/picard/analysis/TheoreticalSensitivityTest.java
… src/tests/java/picard/analysis/directed/CollectHsMetricsTest.java
… src/tests/java/picard/analysis/directed/CollectTargetedMetricsTest.java
… testdata/picard/analysis/TheoreticalSensitivity/Solexa332667_BaseQ.histo
… testdata/picard/analysis/TheoreticalSensitivity/Solexa332667_DepthDist.histo
… testdata/picard/analysis/TheoreticalSensitivity/test_25103070136.targeted_pcr_metrics
… testdata/picard/analysis/TheoreticalSensitivity/test_NexPond-359781.hsMetrics
… testdata/picard/analysis/TheoreticalSensitivity/test_Solexa-316269_sampled.wgs_metrics
… testdata/picard/analysis/TheoreticalSensitivity/test_Solexa-332667.wgs_metrics