Theoretical Het SNP Sensitivity and Het SNP Sensitivity Q Score metrics added to HS metrics, WGS metrics, Targeted PCR metrics, and Sampled WGS metrics (RapidQC). Added tests for each in Theoretical Het Test and added this metric to the respective tests for each collector. 

The length of time added by the calculation of the Theoretical Het Sensitivity metrics is laid out in the table below. These times are from running the unit tests on full sized metrics files in IntelliJ with -Xmx1G as a VM setting. The part that takes the longest is the random sampling. In the standard settings in the code it performs 10,000 samplings, which takes longer with less data in the targeted metrics. The standards in the tests are 1,000 samplings so that the unit tests are not unnecessarily long.
 
Sample Type | Time 
--- | --- 
WGS | 7 seconds
Hybrid Selection | 1 minute 20 seconds
Targeted PCR | 1 minute 2 seconds
WGS Sampled | 9 seconds