@kbergin 2015-12-07
I will add a corresponding flyway pull request.