@kbergin 2015-12-08
@ktibbett @geoffjentry @jacarey this branch is passing all tests locally but the "multiRowFileTest" in ClocsFileReaderTest which is not something I've edited .. I've rebased onto picard and pulled htsjdk. Have you seen this happen before?