@tfenne 2015-12-08
@kbergin I've definitely seen that test fail sporadically in travis before, though I haven't spent the time to track down why.