@kbergin 2015-12-08
@davidbenjamin let me know if you'd like to meet to go through the PR together, it's a seemingly large PR so if it's easier I can help explain all the changes in general before you dig in with the fine tooth comb :)