@davidbenjamin 2015-12-20
@kbergin Looks good, you can squash and merge.