@kbergin 2016-01-12
Further testing of full runs with changes: 

Sample Type | Sample | Original Run Time (min) | Theoretical Het Run Time (min) | Notes 
--- | --- | --- | --- | --- |
WGS Failed at 1G Run on SGE @Xmx2G | Solexa-272222.bam | 386.63 | 408.25 | Actually calculating Het Sens only took 4 seconds, given no additional histograms are being tallied in the new WGS, the time difference is probably due to different loads on the SGE.
HSMetrics on VM @Xmx2G | NexPond-359781.bam | 30.65 | 22.45 | Does the VM cache the bam information for future runs? The bam record processing happened faster the second time. Metric output matches that of current picard output.
Targeted PCR on VM @Xmx1G | 25103070136.bam | 9.39 | 13.40 | Again, the original version was run second and appears to run faster potentially only for that reason. When I ran these two previously, they were only 10 seconds different in time.
WGS Sampled on SGE @Xmx1G | Solexa-316269,_ PCRminus,_ WGSv3,_ 500ng_input,_NA12878.bam | 47.79 | 47.89 |

Note that the aforementioned times in first comment were solely for Theoretical Het Sensitivity calculation, and did not account for the entire time it takes to run the tools. I can supply the full paths to files and run arguments used if needed.

@ktibbett: yossi had suggested seeing if I could borrow two nodes to myself to more accurately test these run times (original on one, test on the other), and that you may know how to help me with that? I'm not worried about the run time, so far it is very similar to the current picard versions, but if you are concerned I can keep testing.

I believe that this tool is ready to be merged after a rebase/squash.