@yfarjoun 2016-01-13
WOOT!

On Wed, Jan 13, 2016 at 4:13 PM, Kylee Bergin <notifications@github.com>
wrote:

> Merged #355 <https://github.com/broadinstitute/picard/pull/355>.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/broadinstitute/picard/pull/355#event-514230193>.
>